Add requirements section
Domain assumptions and requirements are now referenceable with both label and text in every document now, thanks to myitem command and by using \nameref{tag}, so that when modified will now be updated automatically in all positions where they are referenced. This should improve maintainability and reduce errors/discrepancy among the document.